refactor: cardano replace ada folder to cardano since its the correct fullname of the network
