+++ src/cardano/types.ts
@@ -1,8 +1,6 @@
 import type { AvailableWallet } from './available-wallets';
 
-export interface CardanoProviderProps {
-  wallet?: AvailableWallet
-}
+export type CardanoProviderProps = Record<AvailableWallet, any>;
 
 export type CardanoUsedAddress = string
