+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './ada/index'
+export * from './cardano/index'
 export * from './networks'
 export * from './substrate/dot/index'
 export * from './substrate/ksm/index'
