+++ src/types.ts
@@ -1,4 +1,4 @@
-import type { CardanoProviderProps } from './ada/types'
+import type { CardanoProviderProps } from './cardano/types'
 import type { NetworkKey } from './networks'
 import type { SubstrateProviderProps } from './substrate/types'
 
@@ -19,7 +19,7 @@ export type ProviderBuilderProps<T extends NetworkKey> = T extends 'dot' | 'ksm'
 // TODO: Add type and functions of injected providers
 export type Web3Window = {
   injectedWeb3: any
-  cardano: any
+  cardano?: any
 } & Window & typeof globalThis
 
 export const web3Window = (window as Web3Window)
